feedback + follow ups
- extract Path -> path.[ch]pp, privatize members and add a handful
  Is/Has/Get accessors
- add variadic & vector Join funcs (match SQL)
- cleanup/add to the SQL .test path_join tests (correct '' defaulting)
