+++ src/common/CMakeLists.txt
@@ -49,6 +49,7 @@ add_library_unity(
   file_buffer.cpp
   file_system.cpp
   filename_pattern.cpp
+  path.cpp
   fsst.cpp
   gzip_file_system.cpp
   hive_partitioning.cpp
