+++ test/sql/function/string/path_join.test
@@ -1,43 +1,45 @@
 # name: test/sql/function/string/path_join.test
 # group: [string]
 
+# NOTE: far more complete tests available under test/common/test_file_system.cpp ; these largely test SQL specifics
+
 query I
-SELECT replace(path_join('dir//subdir/', './file'), chr(92), '/');
+SELECT path_join('dir//subdir/', './file');
 ----
 dir/subdir/file
 
 query I
-SELECT replace(path_join('dir/subdir', '../sibling'), chr(92), '/');
+SELECT path_join('dir/subdir', '../sibling');
 ----
 dir/sibling
 
 statement error
-SELECT replace(path_join('dir', '/abs/path'), chr(92), '/');
+SELECT path_join('dir', '/abs/path');
 ----
 incompatible paths
 
 query I
-SELECT replace(path_join('dir///', 'nested///child'), chr(92), '/');
+SELECT path_join('dir///', 'nested///child');
 ----
 dir/nested/child
 
 query I
-SELECT replace(path_join('dir', 'subdir', 'file.txt'), chr(92), '/');
+SELECT path_join('dir', 'subdir', 'file.txt');
 ----
 dir/subdir/file.txt
 
 query I
-SELECT replace(path_join('dir/..', '../..'), chr(92), '/');
+SELECT path_join('dir/..', '../..');
 ----
 ../..
 
 query I
-SELECT replace(path_join('/usr/local', '../..'), chr(92), '/');
+SELECT path_join('/usr/local', '../..');
 ----
 /
 
 query I
-SELECT replace(path_join('/', 'usr/local'), chr(92), '/');
+SELECT path_join('/', 'usr/local');
 ----
 /usr/local
 
@@ -81,6 +83,34 @@ SELECT path_join('dir', '/abs/path');
 ----
 incompatible paths
 
+# empty and NULL checks
+
+statement error
+SELECT path_join();
+----
+No function matches
+
+# base relative path normalized to '.'
+query I
+SELECT path_join('');
+----
+.
+
+query I
+SELECT path_join('', '');
+----
+.
+
+query I
+SELECT path_join(NULL);
+----
+NULL
+
+query I
+SELECT path_join(NULL, NULL);
+----
+NULL
+
 query I
 SELECT path_join(NULL, 'dir');
 ----
@@ -95,3 +125,15 @@ query I
 SELECT path_join('base', NULL, 'dir');
 ----
 NULL
+
+# Other misc tests
+query I
+SELECT path_join('asdfoijasdofijasdofijasüîø✅xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy-/QQQQ/', 
+	'IIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIII');
+----
+asdfoijasdofijasdofijasüîø✅xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy-/QQQQ/IIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIII
+
+query I
+SELECT path_join(A, B, C, B, A) FROM (SELECT A:'a', B: 'b', C: 'c');
+----
+a/b/c/b/a
